feature: Clean up threading and job execution

+++ terrible-batch-cleaner/src/main/java/io/terrible/batch/cleaner/processors/CleanProcessor.java
@@ -3,15 +3,17 @@ package io.terrible.batch.cleaner.processors;
 
 import io.terrible.batch.data.domain.MediaFile;
 import io.terrible.batch.data.repository.MediaFileRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.batch.item.ItemProcessor;
+import org.springframework.lang.NonNull;
+
 import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.LinkOption;
 import java.nio.file.Paths;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.FileUtils;
-import org.springframework.batch.item.ItemProcessor;
-import org.springframework.lang.NonNull;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -23,10 +25,12 @@ public class CleanProcessor implements ItemProcessor<MediaFile, MediaFile> {
   public MediaFile process(@NonNull final MediaFile input) {
 
     if (!Files.exists(Paths.get(input.getPath()), LinkOption.NOFOLLOW_LINKS)) {
-
       log.info("Cannot find {} - Removing record", input.getName());
 
-      FileUtils.deleteQuietly(new File(input.getThumbnailPath()));
+      if (StringUtils.isNotBlank(input.getThumbnailPath())) {
+        FileUtils.deleteQuietly(new File(input.getThumbnailPath()));
+      }
+
       mediaFileRepository.delete(input);
     }
